Conflicts:
	rpms/awips2.core/Installer.aviation/component.spec
	rpms/awips2.core/Installer.common-base/component.spec
	rpms/awips2.core/Installer.database/component.spec
	rpms/awips2.core/Installer.gfesuite/component.spec
	rpms/awips2.core/Installer.hydroapps/component.spec
	rpms/awips2.core/deploy.builder/build.sh
	rpms/awips2.edex/Installer.edex-configuration/component.spec
	rpms/awips2.upc/Installer.ldm/component.spec
	rpms/build/common/lookupRPM.sh
	rpms/build/x86_64/build.sh
	rpms/unidata/rsync.dirs

+++ edexOsgi/build.edex/opt/db/replication/setup-standby.sh
@@ -22,9 +22,8 @@
 # Configuration ###############################################################
 
 # Credentials
-db_superuser=awips  # awipsadmin on 16.4.1 and later
+db_superuser=awips
 db_rep_user=replication # for connecting to master
-db_rep_password=replication
 
 # Master server info
 master_hostname="$1" # from command line
@@ -34,6 +33,13 @@ master_port=5432
 this_host=$(hostname -s)
 local_port=5432
 data_dir=/awips2/data
+ssl_dir=/awips2/database/ssl
+
+# For logging the output of this script
+log_dir=/awips2/database/replication/logs
+# Keep this many logs, delete old ones
+keep_logs=5
+log_file="${log_dir}/setup-standby.$(date +%Y%m%d.%H%M%S).log"
 
 # Location of PostgreSQL install
 pg_dir=/awips2/postgresql
@@ -43,11 +49,15 @@ pg_basebackup=${pg_dir}/bin/pg_basebackup
 pg_ctl=${pg_dir}/bin/pg_ctl
 psql=/awips2/psql/bin/psql
 
+log() {
+    echo $* | sudo -u awips tee -a "${log_file}"
+}
+
 ###############################################################################
 
 
 do_pg_ctl() {
-    "${pg_ctl}" -o \"--port=${local_port}\" -D "${data_dir}" $* >/dev/null 2>&1
+    sudo -u awips "${pg_ctl}" -o \"--port=${local_port}\" -D "${data_dir}" $* >/dev/null 2>&1
     return $?
 }
 
@@ -61,21 +71,22 @@ stop_server() {
 
 
 cleanup_exit() {
-    echo "ERROR: There were one or more errors; see above."
-    echo "INFO: Cleaning up."
+    log "ERROR: There were one or more errors; see above."
+    log "INFO: Cleaning up."
     stop_server
     if [[ "$?" -eq 0 ]]; then
-        sleep 1
-        rm -rf "${data_dir}"/*
+        if [[ -d "${data_dir}" ]]; then
+            rm -rf "${data_dir}"/*
+        fi
     else
         # I don't know if this is possible, but if it is, we don't want to
         # delete data dir while server is running
-        echo -n "WARNING: Postgres is still running. "
-        echo "See ${data_dir}/pg_log/postgresql-$(date +%A).log for possible errors."
+        log -n "WARNING: Postgres is still running. "
+        log "See ${data_dir}/pg_log/postgresql-$(date +%A).log for possible errors."
     fi
     if [[ -d "${config_tmpdir}" ]]; then
         if [[ -f "${config_tmpdir}/pg_hba.conf" ]]; then
-            mv "${config_tmpdir}/pg_hba.conf" ${data_dir}
+            sudo -u awips mv "${config_tmpdir}/pg_hba.conf" ${data_dir}
         fi
         if [[ -d "${config_tmpdir}/pg_log" ]]; then
             if [[ -d "${data_dir}/pg_log" ]]; then
@@ -83,8 +94,8 @@ cleanup_exit() {
             else
                 logdir="${data_dir}/pg_log"
             fi
-            echo "INFO: Moving old logs to ${logdir}"
-            mv "${config_tmpdir}/pg_log" "${logdir}"
+            log "INFO: Moving old logs to ${logdir}"
+            sudo -u awips mv "${config_tmpdir}/pg_log" "${logdir}"
         fi
         rm -rf "${config_tmpdir}"
     fi
@@ -101,8 +112,8 @@ if [[ -z "${master_hostname}" ]]; then
     exit 1
 fi
 
-if [[ "$(id -u)" -ne "$(id -u awips)" ]]; then
-    echo "$(basename $0): Must run as user 'awips'."
+if [[ "$(id -u)" -ne 0 ]]; then
+    echo "$(basename $0): Must run as root."
     exit 1
 fi
 
@@ -114,7 +125,6 @@ if [[ "${master_hostname}" == "${this_host}" ||
     echo "Choose a different server name."
     exit 1
 fi
-
 
 # Warning prompt
 echo "You are about to configure this server (${this_host}) as a PostgreSQL"
@@ -137,49 +147,83 @@ fi
 
 # Actually do it ##############################################################
 
+# Make log file for script output
+sudo -u awips mkdir -p "${log_dir}" || exit 1
+sudo -u awips touch "${log_file}" || exit 1
+# Purge old logs
+sudo -u awips find "${log_dir}"/*.log -xdev \
+    | sort \
+    | head -n -${keep_logs} \
+    | tr '\n' '\0' \
+    | sudo xargs -0r rm
+
+log "INFO: Starting replication setup on ${this_host}:${local_port}"
+log "INFO: Will replicate ${master_hostname}:${master_port}"
+
 stop_server || exit 1
 trap 'cleanup_exit' SIGINT
 
-# Backup pg_hba.conf and old logs
-config_tmpdir=$(mktemp -d --tmpdir=${data_dir} .tmp.XXXXXX || cleanup_exit)
+# Get certificates from master
+master_ssl_dir="${ssl_dir}/replication/${master_hostname}"
+sudo -u awips mkdir -p "${master_ssl_dir}"
+log "INFO: Downloading SSL certs and keyfile from ${master_hostname}"
+# must ssh as root to skip password prompt
+rsync --delete-before -av -e ssh \
+    "${master_hostname}":"${master_ssl_dir}"/{replication.crt,replication.key,root.crt} \
+    "${master_ssl_dir}" || exit 1
+chown -R awips:fxalpha "${ssl_dir}"/replication
+find "${ssl_dir}"/replication -xdev -type f -exec chmod 600 {} \;
+find "${ssl_dir}"/replication -xdev -type d -exec chmod 700 {} \;
+
+# Backup pg_hba.conf and old postgres logs
+config_tmpdir=$(sudo -u awips mktemp -d --tmpdir=${data_dir} .tmp.XXXXXX || cleanup_exit)
 if [[ -f "${data_dir}/pg_hba.conf" ]]; then
-    cp -a "${data_dir}/pg_hba.conf" "${config_tmpdir}" || cleanup_exit
+    sudo -u awips cp -a "${data_dir}/pg_hba.conf" "${config_tmpdir}" || cleanup_exit
 fi
 if [[ -d "${data_dir}/pg_log" ]]; then
-    cp -a "${data_dir}/pg_log" "${config_tmpdir}" || cleanup_exit
+    sudo -u awips cp -a "${data_dir}/pg_log" "${config_tmpdir}" || cleanup_exit
 fi
 
 # Prepare data directory
+log "INFO: Recreating ${data_dir}"
 if [[ -d "${data_dir}" ]]; then
     rm -rf "${data_dir}"/*
 else
-    mkdir -p "${data_dir}" || exit 1
-    chmod 700 "${data_dir}" || exit 1
+    sudo -u awips mkdir -p "${data_dir}" || exit 1
+    sudo -u awips chmod 700 "${data_dir}" || exit 1
 fi
 
+# SSL connection string parts
+# needed for basebackup and recovery.conf
+sslmode_part="sslmode=verify-ca"
+sslcert_part="sslcert=${master_ssl_dir}/replication.crt"
+sslkey_part="sslkey=${master_ssl_dir}/replication.key"
+sslrootcert_part="sslrootcert=${master_ssl_dir}/root.crt"
+ssl_part="${sslmode_part} ${sslcert_part} ${sslkey_part} ${sslrootcert_part}"
 
 # pg_basebackup will not write to a non-empty directory
 # so we have to make a temporary one
-data_tmpdir=$(mktemp -d --tmpdir=${data_dir} .tmp.XXXX || cleanup_exit)
+data_tmpdir=$(sudo -u awips mktemp -d --tmpdir=${data_dir} .tmp.XXXX || cleanup_exit)
 # Fetch and install base backup
-echo "INFO: Fetching base backup from ${master_hostname}"
-echo "Enter the password for the '${db_rep_user}' role now, if prompted."
-"${pg_basebackup}" \
+log "INFO: Fetching base backup from ${master_hostname}"
+log "Enter the password for the '${db_rep_user}' role now, if prompted."
+sudo -u awips "${pg_basebackup}" \
     --host="${master_hostname}" \
     --verbose --progress --xlog-method=fetch \
     --username="${db_rep_user}" --format=tar --gzip \
     --port=${master_port} \
+    --db="${ssl_part}" \
     -D "${data_tmpdir}" || cleanup_exit
-mv "${data_tmpdir}"/*.tar.gz "${data_dir}" || cleanup_exit
+sudo -u awips mv "${data_tmpdir}"/*.tar.gz "${data_dir}" || cleanup_exit
 
-echo "INFO: Installing base backup to ${data_dir}"
+log "INFO: Installing base backup to ${data_dir}"
 pushd "${data_dir}" > /dev/null || cleanup_exit
-tar xzf "${data_dir}/base.tar.gz" || cleanup_exit
+sudo -u awips tar xzf "${data_dir}/base.tar.gz" || cleanup_exit
 popd > /dev/null
 rm -f "${data_dir}/base.tar.gz"
 
 # Install tablespaces
-echo INFO: Unpacking tablespaces
+log INFO: Unpacking tablespaces
 # On Postgres 9.5 and later we need to read tablespace_map and create the
 # symlinks ourselves
 if [[ -f "${data_dir}/tablespace_map" ]]; then
@@ -188,7 +232,7 @@ if [[ -f "${data_dir}/tablespace_map" ]]; then
         ts_path="$(echo "$line" | cut -d' ' -f2-)"
         if [[ -n "${ts_num}" && -n "${ts_path}" ]]; then
             rm -f "${data_dir}/pg_tblspc/${ts_num}"
-            ln -sf "${ts_path}" "${data_dir}/pg_tblspc/${ts_num}" || cleanup_exit
+            sudo -u awips ln -sf "${ts_path}" "${data_dir}/pg_tblspc/${ts_num}" || cleanup_exit
         fi
     done < "${data_dir}/tablespace_map"
     rm -f "${data_dir}/tablespace_map"
@@ -197,25 +241,33 @@ fi
 # Now unpack each tar in the right place
 for ts_link in "${data_dir}/pg_tblspc"/*; do
     this_ts=$(readlink "${ts_link}")
-    echo -n "  ${this_ts}..."
+    log -n "  ${this_ts}..."
     tar_name=$(basename "${ts_link}")
     if [[ -d "${this_ts}" ]]; then
         rm -rf "${this_ts}"/*
     else
-        mkdir -p "${this_ts}" || cleanup_exit
+        sudo -u awips mkdir -p "${this_ts}" || cleanup_exit
     fi
     pushd "${this_ts}" > /dev/null
-    tar xzf "${data_dir}/${tar_name}.tar.gz" || cleanup_exit
+    sudo -u awips tar xzf "${data_dir}/${tar_name}.tar.gz" || cleanup_exit
     popd > /dev/null
     rm -f "${data_dir}/${tar_name}.tar.gz"
-    echo done.
+    log done.
 done
 
 # Write recovery.conf
-echo "INFO: Writing ${data_dir}/recovery.conf"
-cat > "${data_dir}/recovery.conf" << EOF || cleanup_exit
+
+host_part="host=${master_hostname}"
+port_part="port=${master_port}"
+user_part="user=${db_rep_user}"
+primary_conninfo="${host_part} ${port_part} ${user_part} ${ssl_part}"
+
+log "INFO: Writing ${data_dir}/recovery.conf"
+rm -f "${data_dir}/recovery.conf"
+sudo -u awips touch "${data_dir}"/recovery.conf
+cat >> "${data_dir}/recovery.conf" << EOF || cleanup_exit
 standby_mode='on'
-primary_conninfo='host=${master_hostname} port=${master_port} user=${db_rep_user} password=${db_rep_password}'
+primary_conninfo='${primary_conninfo}' 
 recovery_target_timeline='latest'
 trigger_file='${data_dir}/promote'
 EOF
@@ -224,35 +276,35 @@ rm -f "${data_dir}/recovery.done"
 
 # Install pg_hba.conf
 if [[ -f "${config_tmpdir}/pg_hba.conf" ]]; then
-    echo "INFO: Installing ${data_dir}/pg_hba.conf"
-    mv "${config_tmpdir}/pg_hba.conf" "${data_dir}"
+    log "INFO: Installing ${data_dir}/pg_hba.conf"
+    sudo -u awips mv "${config_tmpdir}/pg_hba.conf" "${data_dir}"
 fi
 
 # Save old pg_logs
 if [[ -d "${config_tmpdir}/pg_log" ]]; then
     logdir_ts=$(date +%F_%H%M%S)
-    echo "INFO: Moving old logs to ${data_dir}/pg_log-${logdir_ts}"
-    mv "${config_tmpdir}/pg_log" "${data_dir}/pg_log-${logdir_ts}"
+    log "INFO: Moving old logs to ${data_dir}/pg_log-${logdir_ts}"
+    sudo -u awips mv "${config_tmpdir}/pg_log" "${data_dir}/pg_log-${logdir_ts}"
 fi
 
 # Start it up and run test query
-echo "INFO: Starting PostgreSQL"
+log "INFO: Starting PostgreSQL"
 do_pg_ctl start -w || cleanup_exit
 
-echo "INFO: Testing read-only connection to standby"
-is_recovery=$("${psql}" \
+log "INFO: Testing read-only connection to standby"
+is_recovery=$(sudo -u awips "${psql}" \
   -U "${db_superuser}" \
   --port=${local_port} \
   --db=metadata \
   -Aqtc "select pg_is_in_recovery();")
 
 if [[ "${is_recovery}" != "t" ]]; then
-    echo "ERROR: It looks like this server failed to start up properly, or is"
-    echo "ERROR: not in recovery mode."
+    log "ERROR: It looks like this server failed to start up properly, or is"
+    log "ERROR: not in recovery mode."
     cleanup_exit
 fi
 
 rm -rf ${config_tmpdir}
 rm -rf ${data_tmpdir}
 
-echo "INFO: Setup is complete. No errors reported."
+log "INFO: Setup is complete. No errors reported."
